Dynamic code deletion support
This CL makes the following changes:

1) Split the nacl_dyncode_copy syscall into:
nacl_dyncode_create
nacl_dyncode_modify

nacl_dyncode_create is binary compatible with nacl_dyncode_copy and a
stub is left in the SDK code to make it source compatible.

nacl_dyncode_modify is disabled unless NACL_ALLOW_DYNCODE_REPLACEMENT
is set.

2) Added this syscall nacl_dyncode_delete:

This marks a region for deletion (bundle heads to HLTs), then uses a
'wind-down' strategy to wait until all threads enter the service runtime
before allowing reuse.

nacl_dyncode_delete is disabled unless NACL_ALLOW_DYNCODE_REPLACEMENT
is set.

3) No longer use HLT padding to store used/unused state:

A side data-structure is now maintained to store the state of dynamic
text regions.  This needed for:
- Keeping track of used/free (no longer store this state in text memory)
- Making sure dyncode_create/dyncode_delete calls are symmetric
- Keeping track of generation numbers for deletion 'wind-down'


4) Adds thread-safe copying code modification:

This uses atomic writes when it can, but falls back to a slower strategy
based on replacing the first byte of each instruction with HLT during
modification and serializing.

5) Adds caching of mmap results for dynamic code data across calls.

Resulting in a performance boost for small insertions.  The cache is 
limited to a single page.


Review URL: http://codereview.chromium.org/3224008

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@3230 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2