Update nacl_signals to trap signals for all platforms.  In standalone sel_ldr we exit with -signum in all cases, while in chrome, we exit with -signum only for untrusted, and let breakpad deal with the other ones.
Affected issues, breakpad support (win,lin,mac,arm) & crash on windows
bug=http://code.google.com/p/nativeclient/issues/detail?id=204
bug=http://code.google.com/p/nativeclient/issues/detail?id=206
bug=http://code.google.com/p/nativeclient/issues/detail?id=207
bug=http://code.google.com/p/nativeclient/issues/detail?id=208
bug=http://code.google.com/p/nativeclient/issues/detail?id=476
test=nacl_signal_unittest (new), signal_handler (plus lots of others) for sel_ldr

Changes:
  remove linux/generic, replace with posix/windows
  on windows install VectoredExceptionHandler
  on posix install sigactions for all common signals
  Added a linked list of handlers.
  Update all test which use a named exit status to use trusted_<signal>, untrusted_<signal>, or <signal>
    If a trusted sel_ldr, untrusted sel_ldr, or other app exit is expected.
  Modified SConstruct to pass 'arch' and 'subarch' to the command_test.py
  Added a new unit test for nacl_signal_unittest

When a signal comes in, we first check if the signal is from trusted or untrusted code.
If untrusted and x86-32, we restore segment registers for 32b x86 posix.  Win32 will die, so no code there.
We then interate through the linked list of handlers until one of them handles the signal.
If no handler takes the signal, we pass it back to the OS.

For stand alone sel_ldr, we add a generic handler which will exit(-signum) if a signal happens in either
trusted or untrusted code.  For sel_ldr in chrome, we add a handler which exits with -signum for untrusted
code but we return for trusted code.  This will drop back to the OS, and allow breakpad, (if enabled) to
trigger.


Review URL: http://codereview.chromium.org/3605006

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@3432 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2