Fix vulnerability in x86-32 signal handler and re-enable tests
Fix the x86-32 signal handler to correctly restore %gs when a fault
occurs in untrusted code.  Otherwise, the signal handler will probably
crash, or could be vulnerable to exploit.

Fix incorrect setting of %fs to %es's old value (a copy-and-paste
error).

Remove setting of registers on return which is not necessary because
the kernel restores registers from the signal stack frame.

Add back explanatory comments that were removed in r3432.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=1035
TEST=run_signal_handler_test

Review URL: http://codereview.chromium.org/3839006

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@3505 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2