Initial implementation/prototype for a fuzzer that generates nexes from an existing nexe by modifying the phr entries, and then passes the modified nexe to sel_ldr.  Any that cause a crash -- that's not due to abort() -- will be saved in a destination directory.
I'm running it right now on opt-linux-x86-64 using all
nacl-x86-64/staging/*.nexe files as fuzzing templates.  I'll leave
this going in a window and see if anything pops up.  A smarter fuzzing
using boundary condition heuristics would be a good idea, though
if/when we go that route that will be in another CL.  There are
placeholders in this code for selecting the fuzzing algorithm to be
used.

This fuzzer seems to be too slow.  On my desktop, the progress
indicator shows that we're getting fewer than 60 nexes per second (25
dots * 64 runs/dot / 27 seconds), which is abysmal.  This is not
parallel -- i should be able to spawn 4-8 sel_ldr's in parallel
(depending on /proc/cpuinfo), and gain a 4-8x right there -- though
just running the requisite number of instances of the fuzzer is
an easy/trivial way to parallelize.

Even then, I think we may need to integrate the fuzzing into sel_ldr
(or use a separate main function) to achieve higher fuzzing rates, I
think, since the process creation, nexe file I/O, etc overhead is
hurting performance.  However, there are probably several internal
code paths that abort in ways from which it would be difficult for an
integrated fuzzer to recover, plus we'd need to reconstitute the
NaClAppDtor method to clean stuff up so that the integrated fuzzer can
clean up after a failure and spawn another NaClApp object.

Manually tested to see that LOG_FATAL crashes are detected (and
ignored).

We will want to take the saved LOG_FATAL aborts and make the
responsible code paths return errors back up the call stack.  After
that, it will be much easier to more of the fuzzing in-process.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=882 https://tracker.corp.google.com/story/show/307434
TEST=None.  This is a testing tool.  Manually tested.


Review URL: http://codereview.chromium.org/3969002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@3565 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2