Modular build: Track identities of input tarballs and patches
This means that when the input tarballs/patches are changed, it will
cause the build system to re-unpack the source and trigger a rebuild
of the changed modules.

So this change fixes the build system to be properly incremental for
those modules that are stored as tarballs + patches -- currently
binutils, gcc and newlib, although we will switch these to use Git as
the canonical source at some point.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=852
TEST=python test_all.py

Review URL: http://codereview.chromium.org/5351001

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@3771 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2