Scons: Remove redundant "/EHsc" compiler flags
This is MSVC's option for enabling exception handling.

This is already added unconditionally for Windows in SConstruct.
Various files specified it a second time to no benefit.

Reducing the sprinkling of MSVC-specific options is useful for making
NaCl build for Windows with MinGW's gcc.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=1283
TEST=trybots (also tested with MinGW with various other changes applied)

Review URL: http://codereview.chromium.org/6525045

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@4329 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2