Fix exit issue on Vista/Win7Atom
TerminateProcess returns causing "exit" to fall through.   This is causing the buildbots
to go red.  To fix this, We must add a check to verify termination was scheduled, then
loop.  As a seperate CL we should find a way to do this without allowing a return
on an untrusted stack.

BUG= nacl1618
TEST= buildbot
Review URL: http://codereview.chromium.org/6804029

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@4774 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2