Add cross-origin loading of NEXEs for MIME type handlers
We only load NEXEs across origins when rendering non-NaCl MIME
types and when the NEXE is stored in a Chrome extension. That limits this
privilege to only being used when the NEXE is being used as a MIME handler for
another MIME type (e.g., application/pdf).
Review URL: http://codereview.chromium.org/6813070

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@4899 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2