Restructure x86-32 data structures so that they are better organized. Includes:
(1) Make Decoder/Validator states be more like objects with inheritance and
virtual functions (but still in c).

(2) Clean up callers of validator code to be more consistent in the use of
the APIs for the x86-32 validator.

(3) Remove redundant fields/structures by splitting parts of decoder state
that corresponds to instructions into a separate state. This results in
only one copy being generated for decoder states (i.e. aren't duplicated in
the instruction (decoder) buffer).

Note: There are no tests for these changes because (1) the current (functional)
tests should be testing the changes; and (2) Part of the purpose of this change
is to make unit testing easier to introduce.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=1615
Review URL: http://codereview.chromium.org/6833033

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@4927 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2