Reverse RPC channel abstractions and test implementation.
This is a partial change.  This CL establishes and tests the channels,
but does not hook anything up on either end (of any significance).

Added an RPC call on the secure command channel so that the plugin can
ask the service runtime for a new socket address on which reverse
secure channels should be set up.  The plugin immediately spawns a
thread which connects to this socket, with the connecting thread
running an RPC service and, on the service runtime side, the connected
socket is used to make RPC client requests.  These RPC client requests
may (eventually) involve setting up new file name service connections,
where the plugin RPC service thread will spawn additional threads to
connect to the same socket address, but with these threads spawning
file name service RPC service loops after the connection is
established.  Currently the reverse RPCs is just for testing.

The testing framework works okay in the plugin and in sel_universal.
Currently sel_ldr will always tests if the RPC-based controller
(plugin/sel_universal) invokes the new reverse_setup RPC.  This
probably should be moved to a simpler test framework, but soon it will
be system tested.

Moved nacl_simple_service to its own directory.

Added scoped_ptr_refcount to nacl_refcount.h for C++ use.

Refactored sel_ldr_launcher to expose startup sequence.  It's pretty
bare naked right now, and we could unexpose some of the intermediate
steps.

Changed sel_universal to accept additional --rpc_load and
--rpc_services arguments to take advantage of sel_ldr_launcher
refactoring.  This is important since the reverse_setup RPC in sel_ldr
should be invoked before the start_module RPC, but the command channel
must be in place.

Still TODO: refactor plugin to use sel_ldr_launcher (there was a
pre-existing TOD in the code for this already).
Review URL: http://codereview.chromium.org/6712056

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@4976 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2