IRT: Hook up to load the integrated runtime (IRT) library in Chromium
Add a function for Chromium-side code to call to provide a file
descriptor for the IRT library.  NaClMainForChromium() cannot open the
file itself since it is sandboxed.

Until the Chromium-side change is committed, the NaCl code has to be
able to fall back to the non-IRT case in order not to break the
integration bots.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=1595
TEST=
 * check that NaCl continues to work in current versions of Chrome
 * also tested in conjunction with http://codereview.chromium.org/6873133/

Review URL: http://codereview.chromium.org/6878084

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@5024 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2