x86 validator build: Fix some non-determinism in the Gyp build
Previously, some actions which generated header files were
instantiated twice.  It was possible that a compile step could be
running at the same time the second action to generate
nc_opcode_table.h is invoked, which can cause the compiler to fail to
open nc_opcode_table.h.

We change the Gyp file to instantiate the header-generation actions
only once.  We remove the "ncvalidate_gen64" target and do all the
generation in "ncvalidate_gen".

We also reduce the differences between Windows and non-Windows builds.
We simplify it so that nc_opcode_table_32.h and nc_opcode_table_64.h
are always the x86-32 and x86-64 versions.  Previously,
nc_opcode_table.h could be 32-bit or 64-bit on Linux, but was always
32-bit on Windows.

Note that:

 * The 'target_base=="ncvalidate_gen"' clause could now be folded into
   the "ncvalidate_gen" target, but for clarity, I have left that out.

 * As I understand it, nc_opcode_table_32.h is not used by the "old"
   x86-32 validator that we actually use in x86-32 NaCl.  It is part
   of the "new" validator, whose support for x86-32 is experimental
   only.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=1729
TEST=validator_tests in Scons
  + run Gyp build on Windows, test the resulting sel_ldr64.exe on hello_world.nexe

Review URL: http://codereview.chromium.org/6904043

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@5062 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2